Fixed Silly Img
The changes Octo made to SillyIMG were good ideas, but the site broke when checked on mobile. I moved some stuff around and did some messing with classes to make it work how I wanted it to on mobile while still having the much nicer look that Octo established for desktop.